Fix for issue https://github.com/andrii1812/galcon-server/issues/9

+++ src/Model/PlanetUpdate.cs
@@ -3,10 +3,10 @@
     public class PlanetUpdate
     {
         public int ID { get; set; }
-        public double Population { get; set; }
+        public int Population { get; set; }
         public int Owner { get; set; }
 
-        public PlanetUpdate(int id, double population, int owner)
+        public PlanetUpdate(int id, int population, int owner)
         {
             ID = id;
             Population = population;
